Update enc typedef, flow, add G27 P3

+++ Marlin/src/lcd/e3v2/common/encoder.cpp
@@ -135,8 +135,8 @@ EncoderState encoderReceiveAnalyze() {
 
     #if ENABLED(ENCODER_RATE_MULTIPLIER)
       #if ENABLED(ENC_MENU_ITEM)
-        int a = ui.enc_rateA;
-        int b = ui.enc_rateB;
+        uint8_t a = ui.enc_rateA;
+        uint8_t b = ui.enc_rateB;
       #endif
       millis_t ms = millis();
       int32_t encoderMultiplier = 1;
